requirements.txt
Flask>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
APScheduler>=3.10.0

